LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-base-appearance-computed-style-expected.txt

button
SIBLING

FAIL UA styles of base appearance <select>. assert_equals: min-inline-size expected "calc-size(auto, max(size, 24px))" but got "0px"
PASS UA styles of base appearance select::picker-icon.
FAIL UA styles of base appearance ::picker(select) assert_equals: position-try-fallbacks expected "start span-end, end span-start, start span-start" but got "block-start span-inline-end, block-end span-inline-start, block-start span-inline-start"
PASS UA styles of base appearance <option>.
PASS UA styles of base appearance option::checkmark.
PASS UA styles of base appearance <optgroup>.
PASS UA styles of base appearance <legend>.
PASS UA styles of base appearance select <button>.
PASS UA styles of base appearance <option> in <optgroup>.
